% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_profiles_n30439x812.R
\docType{data}
\name{mc_profiles_n30439x812}
\alias{mc_profiles_n30439x812}
\title{Cell Painting CDP data}
\format{A data.frame object with 4959 rows}
\usage{
mc_profiles_n30439x812
}
\description{
Cell Painting CDP data
}
\keyword{datasets}

